requirements.txt
langchain-core
langchainhub
langchain-community
langchain-google-genai
python-dotenv